requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
Mesa==3.5.1
networkx==3.6.1
numpy==2.4.4
pandas==3.0.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
scipy==1.17.1
six==1.17.0
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0

# Config / serialization
PyYAML==6.0.2

# Testing
pytest==9.0.3
pytest-asyncio==1.3.0

# Text-client prototype stack
httpx==0.28.1
textual==8.2.3
websockets==16.0
sseclient-py==1.9.0

# Dev server (graceful reload that emits server_shutdown SSE to clients)
watchfiles==1.1.1
